==> src/analysis/06_make-summary-plot-and-table-parallel.Rmd <==
---
title: "06_make-summary-plot-and-table-parallel"
output: github_document
---

```{r setup, message = FALSE, warning=FALSE}
library(tidyverse)
library(here)
library(ggpubr)
library(kableExtra)
library(magick)
```

Load Data
```{r message=FALSE, warning =FALSE}
mean_prop_completed <- readRDS(here("results", "parallel-bootstrap-resamples.rds"))
parallel_resamples <- as.data.frame(mean_prop_completed)
parallel_means_and_ci <- readRDS(here("results", "parallel-bootstrap-mean-and-ci.rds"))
```

Create plot of (parallel) bootstrap distribution
```{r}
parallel_bootstrap_dist_plot <- parallel_resamples %>% 
  ggplot(aes(x = mean_prop_completed)) +
  geom_histogram(fill = "darkblue") +
  geom_vline(xintercept = parallel_means_and_ci$bootmean) +
  geom_vline(xintercept = parallel_means_and_ci$X2.5., 
             linetype = "dashed") +
  geom_vline(xintercept = parallel_means_and_ci$X97.5., 
             linetype = "dashed") +
  labs(title = "Bootstrap distribution of mean proportion of thru-hikers to \nfinish the Pacific Crest Trail 2013-2019",
       x = "Estimate of Mean Proportion")+
  theme_light()+
  annotate("text", 
           x=0.7, 
           y=7000, 
           label = "Lines represent \nbootstrap mean and\n95% confidence interval")
```

Export plot
```{r}
ggsave("parallel-bootstrap-distribution-plot.png", plot=parallel_bootstrap_dist_plot, path="../results")
```

Create and export table
```{r}
names(parallel_means_and_ci) <- c("Observed Mean", 
                         "Parallel Bootstrap Mean", 
                         "Boostrap Confidence Interval Lower Bound (2.5%)",
                         "Boostrap Confidence Interval Upper Bound (97.5%)")
parallel_means_and_ci %>%
  kbl() %>%
  kable_styling(c("bordered", "condensed"), full_width = FALSE, font_size = 12) %>%
  save_kable(file="../results/parallel-means-and-ci.png")
```
