==> src/analysis/06_timing-experiment.Rmd <==
---
title: "03_timing-experiment"
author: "Ericka Smith"
date: "12/1/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r run-timings}
# timings <- bench::mark(
#   {set.seed(72)
#     func1()
#   },
#   {set.seed(72)
#     func2()
#   }
# )
```


```{r save-output}
# timings_plot <- plot(timings)
# 
# ggsave("timing-experiment-plot.tiff", plot=timings_plot, path="../results")
# timings
# 
# timings %>%
#   kbl() %>%
#   kable_minimal(full_width = F) %>%
#   save_kable(file="../results/timing-experient-table.html")
```
